requirements.txt
torch~=1.11.0
torchvision~=0.12.0
numpy~=1.20.3
albumentations~=1.1.0
tqdm~=4.62.3
pillow~=8.4.0